+++ code/classeine-lib/clsn/ui/HBoxContainer.h
@@ -10,5 +10,7 @@ namespace clsn::ui
     {
     public:
         HBoxContainer();
+
+        void doLayout() override;
     };
 }