+++ code/classeine-lib/clsn/ui/renderers/XYContainerRenderer.cpp
@@ -2,26 +2,6 @@
 
 namespace clsn::ui::renderers
 {
-    void XYContainerRenderer::doLayout(Control& baseControl) const
-    {
-        auto& container = static_cast<XYContainer&>(baseControl);
-
-        const auto visibleCount = container.getVisibleCount();
-        if (visibleCount == 0)
-            return;
-
-        const auto position = container.getActualPosition();
-
-        container.iterate([&position](auto& control, auto& constraint)
-        {
-            if (!control.isVisible())
-                return;
-
-            control.setActualPosition({position.getX() + constraint.getX(), position.getY() + constraint.getY()});
-            control.setActualSize(control.getActualPreferredSize());
-        });
-    }
-
     void XYContainerRenderer::paint(Graphics& graphics,
                                   const Region& region,
                                   const Control& baseControl) const